fix: handle errors appended to body by mermaid
- Removes unused `out` div.
- If the svg appended by mermaid is still on the page then
move it to the correct location and cleanup the wrapper.

For some errors the div + svg added by mermaid is not cleaned up
and an error is thrown [here](https://github.com/mermaid-js/mermaid/blob/develop/packages/mermaid/src/mermaidAPI.ts#L505).

This leads to a list permanent error messages that grows until the
preview is refreshed.